Fixing tests and adding build scripts
Added several scripts to the package json to do various acitions in different order, such as all , which will clean the build directories, run npm install, run the build, and then test the built files.
Removed run mocha with typescript directly as mocha cannot find the type definitions when run this way. Can fix this error later if required, but for now we can test the built javascript files.
Fixed the first unit test to expect a text html file.